+++ tests/integration/src/tests/playwright/shared-test.ts
@@ -24,7 +24,7 @@ export const runTest =
       Effect.provide(PWLive),
       Effect.tapCauseLogPretty,
       Effect.annotateLogs({ thread }),
-      Effect.provide(Logger.pretty),
+      Effect.provide(Logger.prettyWithThread(thread)),
       Effect.runPromise,
     )
   }